backend/requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
requests>=2.31.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0

